Fix client endpoint update when promoting to owner
When a client connection is promoted to be the owner connection,
we were failing to update the existing endpoint on server side.
This results in no clientEndpoint being the owner. Therefore,
when endpoints are removed, none of them will result in
`client disconnected operation` and resources of the client will
leak.

Issue is found when investigating an old report.
https://github.com/hazelcast/hazelcast/issues/8777