Introduce member list join versions
A member list version that denotes when a node joins to the cluster is named "member list join version". It is the version number of the member list in which a node is present for the first time. It is used for identifying nodes in the cluster. A node discovers its "member list join version" when it finalizes its join process and can use the value as its unique integer identifier.

The master node reserves a separate member list join version for each node. The overall solution guarantees uniqueness of the member list join versions during the lifetime of the cluster, including occurrences of network partitioning failures to a certain degree.

Please see the TDD for more info about the solution:

https://hazelcast.atlassian.net/wiki/spaces/EN/pages/209748001/ID+Generator+backed+by+Flake+IDs+Design