Skip migrating partitions when collecting HD query results
The results were collected by iterating through all partitions for which
a partition thread is responsible and for which the node was the owner.
This did not check if the migration was still being migrated and the
index may not have been filled with entries yet.
The fix adds the check which makes the query engine get the results
later with a partition operation specific to that partition.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/1427