Fix lock blocking operations interrupt tests
The tests were relying on the thread blocking on a lock operation to
throw an InterruptedException when the thread is interrupted. This only
happens if the operation is retried since the retry logic will check for
the interrupt flag. To simulate a retry, the
OPERATION_CALL_TIMEOUT_MILLIS was lowered to 5 seconds but this also
lowers the heartbeat timeout as the two are coupled. If there is a slow
environment, the blocking operation will be completed with a heartbeat
timeout. The fix increases the timeout - this will increase the test
duration but also make it more resilient to slow environments.

Fixes: https://github.com/hazelcast/hazelcast/issues/8645