Fix the request used for browsing caches by MC
GetCacheEntry request is used by MC to browse cache entries. Currently,
the EntryProcessor it uses is not serializable and doesn't work for
keys on remote members, causing deserialization failure on the member
when it tries to get the entry from another member.

A new operation is being introduced. Since it's intended for internal
usage only and not needed by client, a new method for triggering the
operation is added to AbstractInternalCacheProxy class directly.

Fixes hazelcast/management-center#603