Make GetCacheEntryRequest serializable
GetCacheEntryRequest is used by MC to browse cache entries. Currently,
the EntryProcessor it uses is not serializable and doesn't work for
keys on remote members, causing deserialization failure on the member
when it tries to get the entry from another member. This commit makes
both the EntryProcessor used and its return type serializable.

Fixes hazelcast/management-center#603