Add Reliable ID generator (#11864)
A cluster-wide unique ID generator. Generated IDs are long primitive 
values and are k-ordered (roughly ordered). This type of generator is 
generally known as Flake ID generator.

The IDs contain timestamp component and a node ID component, which is 
assigned when the member joins the cluster. This allows the IDs to be 
ordered and unique without any coordination between members, which makes 
the generator safe even in split-brain scenario (with some limitations).