Fix for OverlappingFileLockException with NearcachePreloader
Problem caused by synchronization problem with nearcache
creation and destroy inside nearcachemanager.

Scenario that leads to exception as follows
1. A nearcache is put to nearcachemanager and takes file lock
inside mutex lock.
2. Thread1 initates nearcache.destroy. It first removes from
nearcache manager. (did not release the lock yet).
3. Thread2 tries to create a nearcahe. Checks and sees
that nearcache is not avaiable with that name, tries to create
the nearcache. And fails with OverlappingFileLockException when
trying to get lock for nearCachePreLoader.
4. Thread1 releases the lock.

Solution is to make 2. and 4. step under the mutex lock that
is used when creating the nearCache.

Fixes https://github.com/hazelcast/hazelcast/issues/11648