Throw if index out of range in ReadResultSet (#12042)
The return value from `ReadResultSet.getSequence()` was very peculiar:
- it returned `0` for index between `size..maxSize`
- returned `-1` for `index >=maxSize`
- threw `ArrayIndexOutOfBoundsException` for `index < 0`

Since this behavior was undocumented it's hardly been relied upon by anyone (except maybe for incorrect code not failing). This change consolidates the behavior so that `IllegalArgumentException` is thrown for all out-of-range `index` values. This makes it consistent with `ReadResultSet.get()`.