Fix for schedule_thenDisposeLeakTest
After investigation failure reported in #11683, only reason I
found is that test is not able to finish in given time limit.

Timeout is 600000 ms and there is 2.000.000 iterations in the test.
That gives only .3ms if they were running serially. There are
parallelization but it still seems very fragile.

I am increasing the timeout as fix.
fixes #11683