Remove default limit for the number of visible instances on MC
hazelcast.mc.max.visible.instance.count group property is used for
limiting the number of visible instances on Management Center. Its
default value is updated to Integer.MAX_VALUE, effectively removing
the limit if it's not configured explicitly. The current value of 100
confuses user when it's hit, as the users don't configure them. Should
it be needed to limit the number of visible instances, it can still be
configured to a lower number.

Also fixes #12095

(cherry picked from commit 44e7628)