Client authentication retry logic fixed
We introduced a logic in #9506 to prevent a late coming
authentication response to alter client principal
incorrectly. Failure in #11737, showed that using correlation
id as solution is not suitable because correlation id can change
when an invocation is retried.

As a solution, as soon as connection timeout happens, related
future is completed. This way,a late response will not be
processed.

fixes https://github.com/hazelcast/hazelcast/issues/11737