Introduce ICMP config on member side (#12167)
Property based configuration is now deprecated, and will be removed in future versions.
If both properties and configs are in use, the config is taking precedence. Defaults have
been changed to match those of the Client config. However, property defaults stayed the same,
therefore users migrating from 3.9.3 shouldn't see any difference, until the enable the config.

Due to the nature of the implementation (private & internal logic) this was tested manually.