Removed logging of ClassCastException in SplitBrainMergePolicyProvider
This logging causes unwanted noise in case of IMap, ICache and
ReplicatedMap, when using the old merge policies.

This happens since we first try to resolve a new merge policy via the
SplitBrainMergePolicyProvider, which fails when using a legacy merge
policy. Then the data structure specific merge policy provider tries
to resolve the classname.

The logging is superfluous, since the InvalidConfigurationException
includes the original exception.