Fixed ReplicatedMapSplitBrainTest
Since the ReplicatedMap statistics (especially the hit counter) are not
replicated (just migrated), we have to use keys on the same node like
the proxy we are interacting with.

This makes the test less complex, since we can remove the looping over
all proxies, which anyway created unwanted replication operations.