Fix indexed query consistency during migration
There is a short window during partition migration when partition
table is already updated on the destination member, but indexes are
not populated yet. To avoid consulting empty indexes while executing
queries the check for migration status of partitions is performed and
migrating partitions are skipped.

During the last step of query processing skipped partitions are asked
again for the results, see `AbstractMapQueryMessageTask#invokeOnMissingPartitions`.
Usually this works well, but there is a chance that some of the
partitions are still migrating and their indexes are still empty. There
was no check for this situation.

This change brings following:

1. Allows queries to run on a migrating partition if it's a source of
the migration operation since indexes are guaranteed to be available in
this case. This is an optimization.

2. Adds check for migration status of the partition in
`QueryRunner#runPartitionIndexOrPartitionScanQueryOnGivenOwnedPartition`
and fall backs to full scan if partition is a migration destination.
This is the fix.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/1851
See also: https://github.com/hazelcast/hazelcast/issues/8385