Backporting #7865 #7835 and #7635
Fixes are not backported completely. Just the parts that
necessary.

from https://github.com/hazelcast/hazelcast/pull/7865

Fixes hanging authentication future

When connection gets exception from socket and destroyConnection
is called before connection is authenticated, client does not
have an endpoint and it is not in connections map. Because of
that closing that connection was missing. I have added close
for that.

from https://github.com/hazelcast/hazelcast/pull/7835/files

Main issue we are trying to fix is to make sure there are
no invocations left after client has shutdown.

One of the reason that an invocation is left there was
clusterService thread. In ClusterService shutdown we will
wait for executor to shutdown completely to make sure, there
is no invocation left triggered by cluster thread.

It turns out that CleanResources Task logic is indeed not suitable
for using when client is shutting down. Because it was postponing
clearing some invocation for 1 seconds if connection is closed
in last 5 seconds. Much simpler version of clearing invocations
is implemented in place of it.

https://github.com/hazelcast/hazelcast/pull/7635

Fixes a race when connection is closed.

When connection closed, it was possible to endup with invocations
that are not notified. These invocations was causing infinite
wait. We let the race happen and check the callIdMap from another
thread every seconds to cleanup unlucky invocations.