Renamed Inbound/OutboundHandler impls to Decoder/Encoder. (#12232)
They are encoding/decoding Packets/ClientMessage/Text; so they have
been renamed to reflect that. The long names were a not that easy to
understand as the newer snappier names

This renaming follows from the IO pipeline cleanup. With more
handlers, having many classes with name InboundHandler/OutboundHandler
becomes a bit of a pita.