Fix MapListenerTest
When running locally, seems the test sets the
hazelcast.map.entry.filtering.natural.event.types property on some runs
but on others it stays false. This causes the test to fail. Changed
setting the property via System.setProperty to a config property on the
hazelcast config.

Also changed a couple of other tests to prefer hazelcast config
properties or to use OverridePropertyRule.

Fixes: https://github.com/hazelcast/hazelcast/issues/11285