Remove verifying precarious condition in MapLoaderFailoverTest
The removed verification may fail, because the following sequence of events may happen:

1) asyncGetting a particular key on node X
2) Loading entries from store is triggered by 1)
3) GetOperation retries a couple of times on node X, waits for loading to complete
4) Node X gets terminated
5) GetOperation continues on node Y
6) GetOperation completes on node Y and returns null
7) The requested key gets loaded on node Y

There is a race condition between the migrated GetOperation and the restarted loading process. That means the test's expectation was not in sync with the current implementation (loading asynchronously on the nodes) and may fail occasionally.

Fixes #5257