Move ClientJoinOperation to PreJoin
Client ownership info is carried via OnJoinClientOperation.
It was used to work after join which was creating a potential race.
In rare cases, it was possible that a recently joined member is
not aware of the client, if a client is also joined recently.

ClientReAuthOperation is the operation distributes the client
ownership info when a client joined the cluster.

As a fix, the related operation is running before join.
And, ClientReAuthOperation is invoked via invokeOnStableCluster