Fix partition listener in IScheduledFutureProxy
Partition listener was assuming a zero-based index of replicas lost.
Following the IPartitionLostEvent JavaDoc, which clarifies that 0 means owner is lost,
1 that both owner and first backup are lost and so on. However, the PartitionLostEvent,
has a different definition, where it states, that when a partition is completely lost, then
MAX_BACKUP_COUNT is returned. I verified the behavior by looking at the MigrationManager#createPromotionMigrationIfOwnerIsNull