Increase write-behind delay in test
There was a fix back in the day to fix this https://github.com/hazelcast/hazelcast/pull/8599,
which seems to have worked well so far. However, the test is still assuming perfect timing
between instructions which in our build environment is unrealistic.
Especially when we get reported hiccups such as:

> 17:20:35, accumulated pauses: 3846 ms, max pause: 1791 ms, pauses over 1000 ms: 2
> 17:20:40, accumulated pauses: 1865 ms, max pause: 532 ms, pauses over 1000 ms: 0
> 17:20:45, accumulated pauses: 81 ms, max pause: 62 ms, pauses over 1000 ms: 0

I can reproduce the failure by putting a sleep before reading the `timeBeforePut` timestamp.
The reason this fails is that we don't have insight on when the worker thread, responsible
for processing the write-behind ops, will kick in, at what intervals.
So if there is a significant delay before getting the `timeBeforePut` timestamp, and the `put`
then the write-behind worker will kick in sooner than we anticipated.

I can't think of any trivial way to approach this, other than increasing drastically
the `writeDelaySeconds` from 2 secs to 10 secs. An alternative approach could be to keep
track of all flushes and assert their intervals, but this would also be error prone,
i.e., dealing with times.