Make mastership claim idempotent and retry until timeouts or target leaves
Mastership claim is expected to be finalized with one of the following:
- target member explicitly accepts the claim
- target member explicitly rejects the claim
- target member fails
- claim process timeouts

But in a rare case, if a network interruption occurs just around claim is being sent,
claim may be lost over the channel. In this case, sender will wait for an answer
to its claim but target never receives any claim. This leads to claim
process to timeout and eventually target splits from the main cluster.

To prevent this case, mastership claim is modified with a simple change
to make it idempotent, and new claims are sent periodically to target members
within timeout.

Additionally, packet filtering mechanism for testing is improved to distinguish
rejecting and dropping packets. And existing tests changed accordingly.