Retry fetching partition table until target responds or leaves cluster
During mastership change, new master fetches partition tables from the cluster
and determines the latest partition table published by former master.

But in a rare case, if a network interruption occurs just around partition table
request is being sent, request may be lost over the channel.
In this case, new master will wait for an answer forever
but target never receives any request. Or response may be lost and master still waits
forever for an answer.
This can make new master unable to progress
and prevent it to redistribute partitions and to recover missing backups.

To fix this case, partition table fetching operation (which is readonly, hence idempotent)
is sent periodically to all members those are not responded yet.