Retry migration commit when operation timeouts
Migration commit is a blocking operation, which is sent to migration destination
after migration & replication process is completed. Destination should either
explicitly accept/reject the commit or leave the cluster. When it leaves the cluster,
commit is assumed as failed.
That's why migration commit is submitted with an infinite timeout.

But in a rare case, if a network interruption occurs just around migration commit is being sent,
commit may be lost over the channel. In this case, master will wait for an answer forever
but destination never receives the commit. Or response of commit may be lost and still master waits
forever for a response.
This can make new master unable to progress anymore and prevent it to redistribute partitions
and to recover missing backups.

To fix this case, migration commit operation (which is idempotent) is resent to destination
when operation timeouts until destination responds or leaves the cluster.