Retry promotion commit when operation timeouts
Promotion commit is a blocking operation, which is sent to promotion target
after a member is removed from cluster to recover missing partition ownerships.
Target should either explicitly accept/reject the commit or leave the cluster.
When it leaves the cluster, commit is assumed as failed.
That's why promotion commit is submitted with an infinite timeout.

But in a rare case, if a network interruption occurs just around promotion commit is being sent,
commit may be lost over the channel. In this case, master will wait for an answer forever
but target never receives the commit. Or response of commit may be lost and still master waits
forever for a response.
This can make master unable to progress anymore and prevent it to redistribute partitions
and to recover missing backups.

To fix this case, promotion commit operation is improved to make it idempotent
and is resent to target when operation timeouts until target responds or leaves the cluster.