ExecuteScriptRequest/ThreadDumpRequest error reporting (#12437)
* ExecuteScriptRequest/ThreadDumpRequest error reporting

These two classes don't make any difference between an error result and
a success result. With this fix, a distinction is made so that MC can
log the error in its logs and show the success result on the UI.

ExecuteScriptRequest is changed so that it returns responses from each
member in a separate JSON field so that MC can report error/success
from each member separately.

Also, unused fields are removed from ScriptExecutorOperation, while
taking care to preserve its compatibility with previous member versions.
In the forward port of this change, we can it into a version aware
operation.

* Add scriptResult to ExecuteScriptRequest's response

It's needed for compatibility with older MC versions

(cherry picked from commit c40e642)