Remove WAN event list from EP operations
The list was sent from primary to backup replicas for all of the entries
updated or removed by the entry processor. This list is not used at all
and there is no reason why the entry events should be collected or sent.
The WAN events on the backup replica are generated by the
EntryOperator.doPostOperateOps method, just as on the primary and the
list sent from the primary is thrown away.
The change removes this list but keeps backwards compatibility for 3.9.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/1441