Fix situation DONE_ITEM generated a WM (#773)
When processing DONE_ITEM we can generate a WM. The CIES.drainTo added
the WM to inbox, returned MADE_PROGRESS result and a subsequent call to
CIES.isDone() returned `true`. This confused the algorithm.

We fix it by returning DONE result in this situation and by fixing
ProcessorTasklet to handle this situation.