Tolerate event journal reads of stale and future items (#12429)
* Tolerate event journal reads of stale and future items

Previously when the reader requested an event from the event journal
which was already overwritten, the reader would get a
StaleSequenceException. Now we tolerate reading items with stale
sequences and return the oldest events instead, together with a
lostCount field indicating how many events were lost.
Also, allowed the reader to request a future, non existent sequence.
This can happen when some of the partitions have been lost, and it will
try to read from an sequence which doesn't exist. In this case, the read
silently returns from the event following the newest event (i.e. always
blocks and waits for the next event in the journal).
Also, reformatted some javadoc to adhere to a ~72 width limit and
removed some unnecessary rolling upgrade checks.

Fixes: https://github.com/hazelcast/hazelcast/issues/11895