Fixes LoggingScheduledExecutor removeOnCancel problems
In Java 7 there is an efficient removal of cancelled tasks,
so this is enabled regardless of any system property.

In Java 6, there is no efficient removal of cancelled tasks.
So using the group property hazelcast.executionservice.taskscheduler.remove.oncancel
this behavior can be enabled/disabled. This property is now set
to true; even though there can be performance implications due
to the non efficient removal in Java 6. However the disabling
of the property can also cause severe performance problems due
to retention of cancelled tasks.