Fixed merge policy check in ConfigValidator
ConfigValidator.checkMergePolicySupportsInMemoryFormat() used a simple
Class.for() to create a merge policy instance, which didn't work with
simple class names (the default for 3.10 merge policies):

> java.lang.ClassNotFoundException: PutIfAbsentMergePolicy

This PR makes use of the existing merge policy providers to retrieve an
instance of the configured merge policy.

It also replaced CacheProxyUtil.validateCacheConfig() calls with
ConfigValidator.checkCacheConfig(), to prepare the upcoming changes.