[FORWARD PORT] ExecuteScriptRequest/ThreadDumpRequest error reporting (#12595)
* ExecuteScriptRequest/ThreadDumpRequest error reporting

These two classes don't make any difference between an error result and
a success result. With this fix, a distinction is made so that MC can
log the error in its logs and show the success result on the UI.

ExecuteScriptRequest is changed so that it returns responses from each
member in a separate JSON field so that MC can report error/success
from each member separately.

Also, unused fields are removed from ScriptExecutorOperation, while
taking care to preserve its compatibility with previous member versions.
In the forward port of this change, we can it into a version aware
operation.

* Add scriptResult to ExecuteScriptRequest's response

It's needed for compatibility with older MC versions

(cherry picked from commit c40e642)

* ScriptExecutorOperation to read/write int required for compatibility
only if the cluster version is less than 3.10

* Remove scriptResult from ExecuteScriptRequest

It is no longer required for compatibility, because HZ 3.10.x won't
work with MC 3.9.x, we don't need it - MC 3.10 will work with the new
JSON structure

* Make ScriptExecutorOperation implement Versioned

* Use isUnknownOrLessThan for version checks on ScriptExecutorOperation

This is required as ScriptExecutorOperation did not implement Versioned
in 3.9 but implements it in 3.10.