Made LockService a core service
The LockService is an important service, which is used in a lot of
other services. The problem is that NodeEngine.getSharedService()
can return null, so there are null checks all over the place. But
those are not consistent, so we still could run into NPEs.

An example is AbstractRecordStore.createLockStore() which returns
null if the LockService is not available. But later we call
LockStore.getLockedEntryCount() without a null check.

So the code actually relies on the LockService being always
available, but this is not enforced by the current ServiceManager.

This change makes the LockService a core service, so it's always
available before the dependent services are created. This prevents
NPE or null initializations and other races with service
initializations and shutdowns.

The LockService still implements SharedService, so no existing third
party software will break, which makes use of the SPI.