Make OverloadedConnectionsPluginTest more resilient
OverloadedConnectionsPluginTest may fail if the outbound I/O threads can keep up the pace with the load that the only one load producer thread is generating with the async get operations in the test.
Currently, there is no reliable and clean way of throttling outbound connector or queue, therefore this commit attempts to decrease the possibility of test failures by decreasing both the number of I/O threads and the threshold.

Fixes #12713