Make EvictionTest resilient against hiccups
Eviction tests assert that after some time a entry added with TTL is still available in the map. If these tests experience a longer hiccup, an oversleep may happen causing that the entry is no longer available in the map and fails the test. This commit introduces `HazelcastTestSupport#sleepAtMostSeconds`, which addresses hiding these hiccups.

Fixes #12367