Remove time condition from EntryProcessorOffloadableTest
The condition, which asserts that there is a minimal expected time between two heartbeats are sensitive to hiccups. The test's intention is to verify that heartbeats are still coming from the node executing an `Offloadable`. Verifying the number of the expected heartbeats is enough, the time condition can be removed without weakining the guarantees provided by the test.

Fixes: #12744