Fix a race condition while destroying proxies
Before this change there was a race condition while destroying proxies
in a cluster-wide way. Multiple threads were performing destruction of
the same proxy instance simultaneously and none them were able to
succeed at the local resources cleanup.

The idea of the fix is to unify the proxy destruction logic in
ProxyManager to make sure the cluster-wide destruction for a certain
proxy instance is done only once.

Fixes: https://github.com/hazelcast/hazelcast/issues/12679