Use query cache cacheId when registering a preconfigured user listener
Previously the listener registration was done on the cacheName and
the listener would not receive any events, because events are published
with a topic formulated on the basis of cacheId. This fix is only
required for preconfigured user listeners, dynamically added listeners
(via queryCache.addEntryListener) are not affected.

(cherry picked from commit 2d4fde6)