IMap/ICache: Skip replicating records when eviction is required
Normally, we trigger eviction (if needed) after inserting a record. This adds
a small latency (a few ms) to the invocation.

But during migration, many records are inserted (thousands, tens of thousands etc).
If system is close to the eviction threshold and we trigger eviction on each insertion,
this adds a significant latency (a few ms * ten thousand = many seconds) to the migration operation.

Instead of inserting one record and evicting another, we can skip inserting new records
when eviction threshold is reached.