Use entire ringbuffers during split brain healing
Previously single items were provided to the merge policy. The same item
was found by comparing the equality of the items in the ringbuffers.
This is not correct in all scenarios. To provide more flexibility and
allow the user to write a more complex merge policy, we provide the
entire ringbuffers to the user during split brain healing.