Cache config propagation fixes
- CacheGetConfigOperation shouldn't implement
PartitionWideCacheOperation as it is not used as a partition-wide
invocation. Instead, it can be executed on generic operation threads,
as all other cache config related operations.
- In case CacheGetConfigOperation results in registering a new
CacheConfig with the local CacheService, all members of the cluster
should become aware of that config. This is required to guarantee that
the CacheConfig will be available on all members of the cluster
immediately after a Cache has been created. This is implemented
asynchronously so the operation thread doesn't block until all members
have responded.
- CacheGetConfigMessageTask is no longer executed on a particular
partition
- CacheCreateConfigMessageTask now also propagates the cache config
to all members of the cluster asynchronously, while it was previously
blocking the thread it was executing on.