Fixed creation of null MergingValue on merging nodes
The MergingValueFactory had some shortcuts to create null MergingValues
when a given parameter was null or an empty collection. This was meant
to save some code in the merging operations, but could lead to a null
MergingValue on the merging nodes as well. This should not happen, since
our merging policies don't expect the mergingValue to be null (in this
case no merging operation should have been created).