Fixes uninitialized partition in LossToleranceTest
whenLossTolerant_andOwnerCrashes_thenContinue test uses partition service to determine the owner of reliable topic's owner. The owner partition is killed during the test. However owner partition may be null because the partition table was not updated yet. This fix adds warmup stage so that partition owner is guaranteed to be non-null at the time of querying.
Fixes #12882