Close connections to members after they are removed from cluster
When a member is suspected, its connection is closed immediately and if suspecting
member is the master, it removes that member from cluster.
But if a slave member receives a member remove event without suspecting that member,
its connection remains open, unless connection fails with an IO exception.

Additionally, `Invocation` now keeps reference to the connection which operation is sent
through. This is just for logging purposes. Before, `Invocation` was logging
the current/latest connection to the remote target.