Fixes test failures when running with LITTLE_ENDIAN byte order override
- MapIndexLifecycleTest is updated to locate domain objects owned
by the member being asserted for index contents
- BinaryCompatibilityTest runs parametrically in both big & little
endian, so an assumption is added to ensure that the configured byte
order is indeed the expected one. When the byte order override is in
place the assumption is broken and tests are properly ignored.
- StringSerializationTest compares the output of
SerializationService.toBytes with a hand-crafted byte buffer which is
now adapted to match expected contents with LITTLE_ENDIAN byte order.
- AbstractSerializationServiceTest is updated to use the same method
(which takes into account configured byte order) when serializing
values to be compared.

(cherry picked from commit 2b2c2db)