Minor NioThread cleanup
made the class final, so more need for final methods.

Changed the probe level of a bunch of probes from debug
to info. Pretty important what the io threads are doing
and since their number is limited, we won't overwhelm the
system with a huge number of additional probes.

Moved a method to the top.