Fix partition table lazy init bug
When a partition invocation retries because the partition table is not known in a non-master member, the retry occurs on the invocation monitor thread. If the partition table is still not known, PartitionInvocation tries to fetch it from the master by invoking AssignPartitions while initializing the invocation target. However, the internal AssignPartitions invocation fails because the invocation occurs on a OperationHostileThread.

This fix marks AssignPartitions operation as MigrationCycleOperation because it is a cheap one and we don't block the invocation monitor thread for its invocation.