Remove MapLoader from MapIndexLifecycleTest
Instead fill the test maps using plain-old puts. The MapLoader
introduces unnecessary complexity which is unrelated to the purpose of
the test and may result in spurious failures when map loading
operations (eg KeyLoadStatusOperation) are executed on partition
threads concurrently with map destruction on generic operation threads.
This may result in false failures of the test.

(cherry picked from commit d84fbad)