ClientMessage has connection field just like Packet
This makes CLientMessages easier to process. The ClientMessageHandler
and PacketHandler can both be replaced by a Consumer. Which would not
be possible if a connection parameter needs to be passed.

Also the ClientPacket was removed; which was a wrapper around a
client message and a connection. Not needed any longer since
ClientMessage now has connection.