Make EntryProcessorOffloadableTest#testHeartBeatsComingWhenEntryPropcessorOffloaded more resilient against hiccups
If `EntryProcessorOffloadableTest#testHeartBeatsComingWhenEntryPropcessorOffloaded` meets a longer hiccup, the test may observe less heartbeats then what it expects. This fix changes the assertion condition to pass if at least one heartbeats is received while running the test. To enable this verification, the `SimpleValue` inner class is extended by `TimestampedSimpleValue`, which records the start and end timestamps of processing the entry, which can be compared against the recorded heartbeat timestamps.

Fixes #12744
Related #12774