Pulled out population of data adapter from NearCacheTestContext creation
The population of the data adapter in Near Cache tests was done during
the creation of the NearCacheTestContext, because the initial population
sends out invalidation messages and we had no method to properly wait
for these invalidations to be processed. So we had a race between those
invalidations and the following Near Cache population. This lead to a
lot of test failures.

This PR introduces a Near Cache statistic for received Near Cache
invalidations (which are also incremented if no value is removed). This
is needed since the initial invalidation doesn't remove any entries.

The tests have been changed to use the populateDataAdapter() method,
which now waits for those initial invalidations. This resolves the race
condition without mixing the responsibility of the createContext()
method with the test code.

This should also speedup the MapNearCacheBasicTest, since both members
can be created before the data is populated, so there are no more
partition migrations when the second member is started.